requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
eval_type_backport==0.2.2
fastapi==0.115.8
filelock==3.17.0
fsspec==2025.2.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.28.1
idna==3.10
Jinja2==3.1.5
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
langchain-core==0.3.37
langsmith==0.3.8
MarkupSafe==3.0.2
mistralai==1.5.0
mpmath==1.3.0
mypy-extensions==1.0.0
networkx==3.2.1
numpy<2,>=1.26.0
orjson==3.10.15
packaging==24.2
psycopg2-binary==2.9.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==5.2.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.2
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
sympy==1.13.1
tenacity==9.0.0
tokenizers==0.21.0
torch==2.6.0
tqdm==4.67.1
transformers==4.48.3
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
zstandard==0.23.0
